fix formatting of negative numbers

+++ utils.h
@@ -1,4 +1,4 @@
 #include <Arduino.h>
 
 // format a float number to a given precision
-char *toPrecision(char *buffer, long len, float num, byte precision);
+char *toPrecision(char *buffer, size_t len, float num, byte precision);
